Fix initial PR feedback
